Simplify and restructure utilities and docs.
