+++ docs/source/usr_api.rst
@@ -74,7 +74,7 @@ Parameter handling uses the following primary methods.
    usr_parameters
    usr_parallel_beam_model
    usr_cone_beam_model
-   usr_plot_utils
+   usr_utilities
    usr_preprocess
 
 Preprocessing
