+++ docs/source/usr_preprocess.rst
@@ -1,8 +1,8 @@
 .. _PreprocessDocs:
 
-====================
-Preprocess utilities
-====================
+=============
+Preprocessing
+=============
 
 The ``preprocess`` module provides scanner-specific preprocessing and more general preprocessing to compute and correct the sinogram data.
 See `demo_nsi.py <https://github.com/cabouman/mbirjax_applications/tree/main/nsi>`__ in the
@@ -27,7 +27,7 @@ General preprocess functions
 ----------------------------
 
 .. automodule:: mbirjax.preprocess
-   :members: compute_sino_transmission, estimate_background_offset, interpolate_defective_pixels, correct_det_rotation, multi_threshold_otsu, export_recon_to_hdf5
+   :members: compute_sino_transmission, estimate_background_offset, interpolate_defective_pixels, correct_det_rotation, multi_threshold_otsu
    :undoc-members:
    :show-inheritance:
 
@@ -40,4 +40,4 @@ General preprocess functions
       interpolate_defective_pixels
       correct_det_rotation
       multi_threshold_otsu
-      export_recon_to_hdf5
+
