+++ docs/source/usr_utilities.rst
@@ -1,0 +1,26 @@
+===============
+Plotting and IO
+===============
+
+MBIRJAX includes some utilities to view 3D data slice-by-slice and to save and load recons via hdf5 files.
+
+Here is an example showing views of a modified Shepp-Logan phantom, with changing intensity window and displayed slice.
+
+.. image:: https://www.math.purdue.edu/~buzzard/images/slice_viewer_demo.gif
+   :alt: An animated image of the slice viewer.
+
+Display Functions
+-----------------
+
+.. automethod::
+   mbirjax.utilities.slice_viewer
+
+IO Functions
+------------
+
+.. automethod::
+      mbirjax.utilities.export_recon_to_hdf5
+
+.. automethod::
+      mbirjax.utilities.import_recon_from_hdf5
+
