+++ mbirjax/__init__.py
@@ -6,5 +6,5 @@ from .parallel_beam import *
 from .cone_beam import *
 from .vcd_utils import *
 from .memory_stats import *
-from .plot_utils import *
+from .utilities import *
 from . import preprocess
